+++ examples/agent-engine-nightly-example/variables.tf
@@ -14,16 +14,7 @@
  * limitations under the License.
  */
 
-terraform {
-  required_providers {
-    google = {
-      source  = "hashicorp/google"
-      version = ">= 7.19, < 8"
-    }
-    google-beta = {
-      source  = "hashicorp/google-beta"
-      version = ">= 7.19, < 8"
-    }
-  }
-  required_version = ">= 1.3"
+variable "project_id" {
+  description = "The ID of the Google Cloud project in which to deploy the Reasoning Engine."
+  type        = string
 }