+++ CHANGELOG.md
@@ -8,6 +8,31 @@ and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 This changelog is generated automatically based on [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
 
+## [5.1.0](https://github.com/GoogleCloudPlatform/terraform-google-vertex-ai/compare/v5.0.0...v5.1.0) (2026-04-10)
+
+
+### Features
+
+* added agent-engine module using nightly provider ([#97](https://github.com/GoogleCloudPlatform/terraform-google-vertex-ai/issues/97)) ([3d4c673](https://github.com/GoogleCloudPlatform/terraform-google-vertex-ai/commit/3d4c67315f723a50548de76de1f5b1a180fc424c))
+
+## [5.0.0](https://github.com/GoogleCloudPlatform/terraform-google-vertex-ai/compare/v4.0.1...v5.0.0) (2026-04-08)
+
+
+### ⚠ BREAKING CHANGES
+
+* added output fields and identity_type input field ([#94](https://github.com/GoogleCloudPlatform/terraform-google-vertex-ai/issues/94))
+
+### Bug Fixes
+
+* added output fields and identity_type input field ([#94](https://github.com/GoogleCloudPlatform/terraform-google-vertex-ai/issues/94)) ([71c41ae](https://github.com/GoogleCloudPlatform/terraform-google-vertex-ai/commit/71c41aeb1ea5abac4aca4758974a7797cae47cc9))
+
+## [4.0.1](https://github.com/GoogleCloudPlatform/terraform-google-vertex-ai/compare/v4.0.0...v4.0.1) (2026-03-20)
+
+
+### Bug Fixes
+
+* added visibility for reasoning engine id ([#91](https://github.com/GoogleCloudPlatform/terraform-google-vertex-ai/issues/91)) ([9934598](https://github.com/GoogleCloudPlatform/terraform-google-vertex-ai/commit/99345986f0fd9b1488e59d02679a803ad15bec71))
+
 ## [4.0.0](https://github.com/GoogleCloudPlatform/terraform-google-vertex-ai/compare/v3.1.1...v4.0.0) (2026-03-03)